+++ packages/yourloops/components/dashboard-widgets/patient-dashboard.tsx
@@ -29,10 +29,8 @@ import React, { type FunctionComponent } from 'react'
 import { type Patient } from '../../lib/patient/models/patient.model'
 import DeviceUsage from 'blip/app/components/chart/deviceUsage.js'
 import type MedicalDataService from 'medical-domain'
-import { DatumType, type TimePrefs } from 'medical-domain'
+import { type TimePrefs } from 'medical-domain'
 import { type BgPrefs } from 'dumb'
-import type DataUtil from 'tidepool-viz/src/utils/data'
-import moment from 'moment-timezone'
 import Grid from '@mui/material/Grid'
 import AccessTime from '@mui/icons-material/AccessTime'
 import { useTranslation } from 'react-i18next'
@@ -45,7 +43,6 @@ import {
   RESPONSIVE_GRID_HALF_WIDTH
 } from '../../css/css-utils'
 import { PatientStatisticsWidget } from './patient-statistics-widget'
-import Stats from 'blip/app/components/chart/stats'
 import MedicalFilesWidget from './medical-files/medical-files-widget'
 import MonitoringAlertsCard from '../monitoring-alert/monitoring-alerts-card'
 import { makeStyles } from 'tss-react/mui'
@@ -56,7 +53,6 @@ import { PRIVATE_TEAM_ID, useTeam } from '../../lib/team/team.hook'
 interface PatientDashboardProps {
   bgPrefs: BgPrefs
   dashboardEpochDate: number
-  dataUtil: typeof DataUtil
   goToDailySpecificDate: (date: number | Date) => void
   loading: boolean
   medicalDataService: MedicalDataService
@@ -77,9 +73,7 @@ export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props
   const {
     bgPrefs,
     dashboardEpochDate,
-    dataUtil,
     goToDailySpecificDate,
-    loading,
     medicalDataService,
     msRange,
     patient,
@@ -93,10 +87,6 @@ export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props
   const { classes, theme } = useStyle()
 
   const isMobileBreakpoint: boolean = useMediaQuery(theme.breakpoints.only('xs'))
-  const endpoints = [
-    moment.utc(dashboardEpochDate - msRange).toISOString(), // start
-    moment.utc(dashboardEpochDate).toISOString() // end
-  ]
   const dateFilter = {
     start: dashboardEpochDate - msRange,
     end: dashboardEpochDate
@@ -144,19 +134,8 @@ export const PatientDashboard: FunctionComponent<PatientDashboardProps> = (props
         <PatientStatisticsWidget
           medicalData={medicalData}
           bgPrefs={bgPrefs}
-          bgType={dataUtil.bgSource}
           dateFilter={dateFilter}
-        >
-          <Stats
-            bgPrefs={bgPrefs}
-            bgSource={DatumType.Cbg}
-            chartType="patientStatistics"
-            loading={loading}
-            dataUtil={dataUtil}
-            endpoints={endpoints}
-            parametersConfig={medicalData?.pumpSettings[0]?.payload?.parameters}
-          />
-        </PatientStatisticsWidget>
+        />
       </Grid>
 
       <Grid item xs={gridWidgetSize}>